REF - Refactor action types into single location to be included in all consumers, to improve on organization and generated documentation.
