+++ app/action_types/actionTypes.js
@@ -1,0 +1,4 @@
+export const SUBMIT_HINT = 'SUBMIT_HINT';
+export const PLAYER_JOIN_GAME = 'PLAYER_JOIN_GAME';
+export const JOIN_GAME = 'JOIN_GAME';
+export const SUBMIT_GUESS = 'SUBMIT_GUESS';
