+++ app/actions/dealer.js
@@ -1,4 +1,4 @@
-export const SUBMIT_HINT = 'SUBMIT_HINT';
+import { SUBMIT_HINT } from '../action_types/actionTypes.js';
 
 export const submitHint = (hint) => ({
   type: SUBMIT_HINT,
